+++ src/controllers/page.js
@@ -4,13 +4,13 @@ import MovieList, {MovieListType} from "../components/movie-list";
 import ShowMoreButton from "../components/show-more-button";
 import MovieBoard from "../components/movie-board";
 import Statistics from "../components/statistics";
-import Sort, {getSortedMoviesBySortType, SortType} from "../components/sort";
+import Sort from "../components/sort";
 import Filter from "../components/filter";
 
 import {remove, componentRender} from "../utils/render";
-import {increaseInt} from "../utils/common";
+import {increaseInt, getSortedMoviesBySortType} from "../utils/common";
 
-import {KEY_CODE, MOVIE_COUNT} from "../constatnts";
+import {KEY_CODE, MOVIE_COUNT, SortType} from "../constants";
 
 /**
  * Поиск фильмов с наивысшими оценками
